mv pr fixt from wh receiver to tests/unit/conf and mv commit part to separate func
